# bdc_stac/spec/openapi.yaml
openapi: 3.0.2
info:
  title: 'STAC API'
  version: 1.0.0-rc.1
  description: This is an OpenAPI definition of the SpatioTemporal Asset Catalog API
    Item Search specification.
  contact:
    name: STAC Specification
    url: http://stacspec.org
  license:
    name: Apache License 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
paths:
...
  /conformance:
    get:
      tags:
      - Features
      summary: information about specifications that this API conforms to
      description: |-
        A list of all conformance classes specified in a standard that the
        server conforms to.
      operationId: getConformanceDeclaration
      responses:
        '200':
          $ref: '#/components/responses/ConformanceDeclaration'
        '500':
          $ref: '#/components/responses/ServerError'
  /collections:
    get:
      tags:
      - Features
      summary: the feature collections in the dataset
      description: |-
        A body of Feature Collections that belong or are used together with additional links.
        Request may not return the full set of metadata per Feature Collection.
      operationId: getCollections
      responses:
        '200':
          $ref: '#/components/responses/Collections'
        '500':
          $ref: '#/components/responses/ServerError'
  /collections/{collectionId}:
    get:
      tags:
      - Features
      summary: describe the feature collection with id `collectionId`
      description: |-
        A single Feature Collection for the given if `collectionId`.
        Request this endpoint to get a full list of metadata for the Feature Collection.
      operationId: describeCollection
      parameters:
      - $ref: '#/components/parameters/collectionId'
      responses:
        '200':
          $ref: '#/components/responses/Collection'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /collections/{collectionId}/items:
    get:
      tags:
      - Features
      summary: fetch features
      description: |-
        Fetch features of the feature collection with id `collectionId`.

        Every feature in a dataset belongs to a collection. A dataset may
        consist of multiple feature collections. A feature collection is often a
        collection of features of a similar type, based on a common schema.
      operationId: getFeatures
      parameters:
      - $ref: '#/components/parameters/collectionId'
      - $ref: '#/components/parameters/limit'
      - $ref: '#/components/parameters/bbox'
      - $ref: '#/components/parameters/datetime'
      responses:
        '200':
          $ref: '#/components/responses/Features'
        '400':
          $ref: '#/components/responses/InvalidParameter'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /collections/{collectionId}/items/{featureId}:
    get:
      tags:
      - Features
      summary: fetch a single feature
      description: |-
        Fetch the feature with id `featureId` in the feature collection
        with id `collectionId`.
      operationId: getFeature
      parameters:
      - $ref: '#/components/parameters/collectionId'
      - $ref: '#/components/parameters/featureId'
      responses:
        '200':
          $ref: '#/components/responses/Feature'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
components:
  schemas:
...
    stac_version:
      title: STAC version
      type: string
      example: 1.0.0-rc.1
    stac_extensions:
      title: STAC extensions
      type: array
      uniqueItems: true
      items:
        anyOf:
        - title: Reference to a JSON Schema
          type: string
          format: uri
        - title: Reference to a core extension
          type: string
...
  responses:
...
    ConformanceDeclaration:
      description: |-
        The URIs of all conformance classes supported by the server.

        To support "generic" clients that want to access multiple
        OGC API Features implementations - and not "just" a specific
        API / server, the server declares the conformance
        classes it implements and conforms to.
      content:
        application/json:
          schema:
            type: object
            required:
            - conformsTo
            properties:
              conformsTo:
                description: A list of all conformance classes implemented by the
                  server. In addition to the STAC-specific conformance classes, all
                  OGC-related conformance classes listed at `GET /conformances` must
                  be listed here. This entry should mirror what `GET /conformances`
                  returns, if implemented.
                type: array
                items:
                  type: string
          example:
            conformsTo:
            - http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/core
            - http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/oas30
            - http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/geojson
    Collections:
      description: |-
        The feature collections shared by this API.

        The dataset is organized as one or more feature collections. This resource
        provides information about and access to the collections.

        The response contains the list of collections. For each collection, a link
        to the items in the collection (path `/collections/{collectionId}/items`,
        link relation `items`) as well as key information about the collection.
        This information includes:

        * A local identifier for the collection that is unique for the dataset;
        * A list of coordinate reference systems (CRS) in which geometries may be returned by the server. The first CRS is the default coordinate reference system (the default is always WGS 84 with axis order longitude/latitude);
        * An optional title and description for the collection;
        * An optional extent that can be used to provide an indication of the spatial and temporal extent of the collection - typically derived from the data;
        * An optional indicator about the type of the items in the collection (the default value, if the indicator is not provided, is 'feature').
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/collections'
    Collection:
      description: |-
        Information about the feature collection with id `collectionId`.

        The response contains a link to the items in the collection
        (path `/collections/{collectionId}/items`, link relation `items`)
        as well as key information about the collection. This information
        includes:

        * A local identifier for the collection that is unique for the dataset;
        * A list of coordinate reference systems (CRS) in which geometries may be returned by the server. The first CRS is the default coordinate reference system (the default is always WGS 84 with axis order longitude/latitude);
        * An optional title and description for the collection;
        * An optional extent that can be used to provide an indication of the spatial and temporal extent of the collection - typically derived from the data;
        * An optional indicator about the type of the items in the collection (the default value, if the indicator is not provided, is 'feature').
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/collections/properties/collections/items'
    Features:
      description: |-
        The response is a document consisting of features in the collection.
        The features included in the response are determined by the server
        based on the query parameters of the request. To support access to
        larger collections without overloading the client, the API supports
        paged access with links to the next page, if more features are selected
        that the page size.

        The `bbox` and `datetime` parameter can be used to select only a
        subset of the features in the collection (the features that are in the
        bounding box or time interval). The `bbox` parameter matches all features
        in the collection that are not associated with a location, too. The
        `datetime` parameter matches all features in the collection that are
        not associated with a time stamp or interval, too.

        The `limit` parameter may be used to control the subset of the
        selected features that should be returned in the response, the page size.
        Each page may include information about the number of selected and
        returned features (`numberMatched` and `numberReturned`) as well as
        links to support paging (link relation `next`).
      content:
        application/geo+json:
          schema:
            $ref: '#/components/schemas/featureCollectionGeoJSON'
    Feature:
      description: |-
        fetch the feature with id `featureId` in the feature collection
        with id `collectionId`
      content:
        application/geo+json:
          schema:
            description: A GeoJSON Feature augmented with foreign members that contain
              values relevant to a STAC entity
            type: object
            required:
            - stac_version
            - id
            - type
            - geometry
            - bbox
            - links
            - properties
            - assets
            properties:
              stac_version:
                $ref: '#/components/schemas/stac_version'
              stac_extensions:
                $ref: '#/components/schemas/stac_extensions'
              id:
                type: string
                description: Provider identifier, a unique ID.
              bbox:
                description: |-
                  Only features that have a geometry that intersects the bounding box are
                  selected. The bounding box is provided as four or six numbers,
                  depending on whether the coordinate reference system includes a
                  vertical axis (elevation or depth):

                  * Lower left corner, coordinate axis 1
                  * Lower left corner, coordinate axis 2
                  * Lower left corner, coordinate axis 3 (optional)
                  * Upper right corner, coordinate axis 1
                  * Upper right corner, coordinate axis 2
                  * Upper right corner, coordinate axis 3 (optional)

                  The coordinate reference system of the values is WGS84
                  longitude/latitude (http://www.opengis.net/def/crs/OGC/1.3/CRS84) unless
                  a different coordinate reference system is specified in the parameter
                  `bbox-crs`.

                  For WGS84 longitude/latitude the values are in most cases the sequence
                  of minimum longitude, minimum latitude, maximum longitude and maximum
                  latitude. However, in cases where the box spans the antimeridian the
                  first value (west-most box edge) is larger than the third value
                  (east-most box edge).

                  If a feature has multiple spatial geometry properties, it is the
                  decision of the server whether only a single spatial geometry property
                  is used to determine the extent or all relevant geometries.

                  Example: The bounding box of the New Zealand Exclusive Economic Zone in
                  WGS 84 (from 160.6°E to 170°W and from 55.95°S to 25.89°S) would be
                  represented in JSON as `[160.6, -55.95, -170, -25.89]` and in a query as
                  `bbox=160.6,-55.95,-170,-25.89`.
                type: array
                minItems: 4
                maxItems: 6
                items:
                  type: number
                example:
                - -110
                - 39.5
                - -105
                - 40.5
              geometry:
                oneOf:
                - type: object
                  required:
                  - type
                  - coordinates
                  properties:
                    type:
                      type: string
                      enum:
                      - Point
                    coordinates:
                      type: array
                      minItems: 2
                      items:
                        type: number
                - type: object
                  required:
                  - type
                  - coordinates
                  properties:
                    type:
                      type: string
                      enum:
                      - MultiPoint
                    coordinates:
                      type: array
                      items:
                        type: array
                        minItems: 2
                        items:
                          type: number
                - type: object
                  required:
                  - type
                  - coordinates
                  properties:
                    type:
                      type: string
                      enum:
                      - LineString
                    coordinates:
                      type: array
                      minItems: 2
                      items:
                        type: array
                        minItems: 2
                        items:
                          type: number
                - type: object
                  required:
                  - type
                  - coordinates
                  properties:
                    type:
                      type: string
                      enum:
                      - MultiLineString
                    coordinates:
                      type: array
                      items:
                        type: array
                        minItems: 2
                        items:
                          type: array
                          minItems: 2
                          items:
                            type: number
                - type: object
                  required:
                  - type
                  - coordinates
                  properties:
                    type:
                      type: string
                      enum:
                      - Polygon
                    coordinates:
                      type: array
                      items:
                        type: array
                        minItems: 4
                        items:
                          type: array
                          minItems: 2
                          items:
                            type: number
                - type: object
                  required:
                  - type
                  - coordinates
                  properties:
                    type:
                      type: string
                      enum:
                      - MultiPolygon
                    coordinates:
                      type: array
                      items:
                        type: array
                        items:
                          type: array
                          minItems: 4
                          items:
                            type: array
                            minItems: 2
                            items:
                              type: number
                - type: object
                  required:
                  - type
                  - geometries
                  properties:
                    type:
                      type: string
                      enum:
                      - GeometryCollection
                    geometries:
                      type: array
                      items:
                        $ref: '#/components/schemas/intersectsFilter/properties/intersects'
              type:
                type: string
                description: The GeoJSON type
                enum:
                - Feature
              links:
                $ref: '#/components/schemas/collections/properties/links'
              assets:
                type: object
                additionalProperties:
                  type: object
                  required:
                  - href
                  properties:
                    href:
                      type: string
                      format: url
                      description: Link to the asset object
                      example: http://cool-sat.com/catalog/collections/cs/items/CS3-20160503_132130_04/thumb.png
                    title:
                      type: string
                      description: Displayed title
                      example: Thumbnail
                    description:
                      type: string
                      description: |-
                        Multi-line description to explain the asset.

                        [CommonMark 0.29](http://commonmark.org/) syntax MAY be used for rich text representation.
                      example: Small 256x256px PNG thumbnail for a preview.
                    type:
                      type: string
                      description: Media type of the asset
                      example: image/png
                    roles:
                      type: array
                      items:
                        type: string
                      description: Purposes of the asset
                      example:
                      - thumbnail
            example:
              stac_version: 1.0.0
              stac_extensions:
              - eo
              - view
              - https://example.com/cs-extension/1.0/schema.json
              type: Feature
              id: CS3-20160503_132131_05
              bbox:
              - -122.59750209
              - 37.48803556
              - -122.2880486
              - 37.613537207
              geometry:
                type: Polygon
                coordinates:
                - - - -122.308150179
                    - 37.488035566
                  - - -122.597502109
                    - 37.538869539
                  - - -122.576687533
                    - 37.613537207
                  - - -122.2880486
                    - 37.562818007
                  - - -122.308150179
                    - 37.488035566
              properties:
                datetime: '2016-05-03T13:22:30.040Z'
                title: A CS3 item
                license: PDDL-1.0
                providers:
                - name: CoolSat
                  roles:
                  - producer
                  - licensor
                  url: https://cool-sat.com/
                view:sun_azimuth: 168.7
                eo:cloud_cover: 0.12
                view:off_nadir: 1.4
                platform: coolsat2
                instruments:
                - cool_sensor_v1
                eo:bands: []
                view:sun_elevation: 33.4
                eo:gsd: 0.512
              collection: CS3
              links:
              - rel: self
                href: http://cool-sat.com/collections/CS3/items/20160503_132130_04
              - rel: root
                href: http://cool-sat.com/collections
              - rel: parent
                href: http://cool-sat.com/collections/CS3
              - rel: collection
                href: http://cool-sat.com/collections/CS3
              assets:
                analytic:
                  href: http://cool-sat.com/static-catalog/CS3/20160503_132130_04/analytic.tif
                  title: 4-Band Analytic
                thumbnail:
                  href: http://cool-sat.com/static-catalog/CS3/20160503_132130_04/thumbnail.png
                  title: Thumbnail
    InvalidParameter:
      description: A query parameter has an invalid value.
      content:
        application/json:
          schema:
            $ref: '#/components/responses/ServerError/content/application~1json/schema'
    NotFound:
      description: The requested URI was not found.
    ServerError:
      description: A server error occurred.
      content:
        application/json:
          schema:
            type: object
            description: 'Information about the exception: an error code plus an optional
              description.'
            required:
            - code
            properties:
              code:
                type: string
              description:
                type: string
  parameters:
...
    datetime:
      name: datetime
      in: query
      description: |-
        Either a date-time or an interval, open or closed. Date and time expressions
        adhere to RFC 3339. Open intervals are expressed using double-dots.

        Examples:

        * A date-time: "2018-02-12T23:20:50Z"
        * A closed interval: "2018-02-12T00:00:00Z/2018-03-18T12:31:12Z"
        * Open intervals: "2018-02-12T00:00:00Z/.." or "../2018-03-18T12:31:12Z"

        Only features that have a temporal property that intersects the value of
        `datetime` are selected.

        If a feature has multiple temporal properties, it is the decision of the
        server whether only a single temporal property is used to determine
        the extent or all relevant temporal properties.
      required: false
      schema:
        type: string
      style: form
      explode: false
    bbox:
      name: bbox
      in: query
      description: |-
        Only features that have a geometry that intersects the bounding box are selected.
        The bounding box is provided as four or six numbers, depending on whether the
        coordinate reference system includes a vertical axis (height or depth):

        * Lower left corner, coordinate axis 1
        * Lower left corner, coordinate axis 2
        * Minimum value, coordinate axis 3 (optional)
        * Upper right corner, coordinate axis 1
        * Upper right corner, coordinate axis 2
        * Maximum value, coordinate axis 3 (optional)

        The coordinate reference system of the values is WGS 84 longitude/latitude
        (http://www.opengis.net/def/crs/OGC/1.3/CRS84).

        For WGS 84 longitude/latitude the values are in most cases the sequence of
        minimum longitude, minimum latitude, maximum longitude and maximum latitude.
        However, in cases where the box spans the antimeridian the first value
        (west-most box edge) is larger than the third value (east-most box edge).

        If the vertical axis is included, the third and the sixth number are
        the bottom and the top of the 3-dimensional bounding box.

        If a feature has multiple spatial geometry properties, it is the decision of the
        server whether only a single spatial geometry property is used to determine
        the extent or all relevant geometries.
      required: false
      schema:
        type: array
        oneOf:
        - minItems: 4
          maxItems: 4
        - minItems: 6
          maxItems: 6
        items:
          type: number
      style: form
      explode: false
    limit:
      name: limit
      in: query
      description: |-
        The optional limit parameter limits the number of items that are presented in the response document.

        Only items are counted that are on the first level of the collection in the response document.
        Nested objects contained within the explicitly requested items shall not be counted.

        Minimum = 1. Maximum = 10000. Default = 10.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 10000
        default: 10
      style: form
      explode: false
    collectionId:
      name: collectionId
      in: path
      description: local identifier of a collection
      required: true
      schema:
        type: string
    featureId:
      name: featureId
      in: path
      description: local identifier of a feature
      required: true
      schema:
        type: string
...